separate electrum interface to its own file
